+++ package.json
@@ -17,8 +17,10 @@
     "prettier:write": "prettier --write {src,test,mocks}/**/*.{ts,tsx,js,jsx}",
     "lint": "tslint '{src,test,mocks}/**/*.{ts,tsx,js,jsx}' --project ./tsconfig.json",
     "lint:fix": "tslint '{src,test,mocks}/**/*.{ts,tsx}' --project ./tsconfig.json --fix",
-    "test": "jest '(\\/test\\/).*'",
-    "test:integration": "jest '(\\/test\\/integration/).*'",
+    "test": "npm run test:unit",
+    "test:unit": "jest --testRegex '\\.test\\.tsx?$'",
+    "test:unit:watch": "jest --testRegex '\\.test\\.tsx?$' --watch",
+    "test:integration": "jest --testRegex '\\.itest\\.ts$'",
     "release:check": "npm run lint && npm test",
     "release": "npm run release:check && npm run build && electron-builder --publish onTag",
     "postinstall": "electron-builder install-app-deps"
@@ -27,7 +29,7 @@
     "transform": {
       "^.+\\.tsx?$": "ts-jest"
     },
-    "testRegex": "(/test/.+\\.spec)\\.tsx?$",
+    "testRegex": "\\.?test\\.tsx?$",
     "moduleFileExtensions": [
       "ts",
       "tsx",
